refactor: OverviewDiv 컴포넌트화 #32
