+++ frontend/src/components/Chart/Overview/OverviewCard.tsx
@@ -1,6 +1,6 @@
-import * as S from './Chart.style';
-import ToolTip from '../common/ToolTip/ToolTip';
-import useCountNum from '../../hooks/useCountUp';
+import * as S from '../Chart.style';
+import ToolTip from '../../common/ToolTip/ToolTip';
+import useCountUp from '../../../hooks/useCountUp';
 
 interface OverviewCardType {
   title: string;
@@ -11,10 +11,10 @@ interface OverviewCardType {
 const OverviewCard = ({ title, count, tooltipText, barColor }: OverviewCardType) => {
   return (
     <S.OverviewCardLayout>
-      <S.VerticalLine barColor={barColor}/>
+      <S.VerticalLine barColor={barColor} />
       <S.InfoDiv>
         <S.TitleP>{title}</S.TitleP>
-        <S.CountP barColor={barColor}>{isNaN(parseInt(count)) ? count : useCountNum(parseInt(count))}</S.CountP>
+        <S.CountP barColor={barColor}>{isNaN(parseInt(count)) ? count : useCountUp(parseInt(count))}</S.CountP>
       </S.InfoDiv>
       <S.ToolTipDiv>
         <ToolTip>
