+++ frontend/src/components/Chart/Overview/OverviewLargeCard.tsx
@@ -1,5 +1,5 @@
-import * as S from './Chart.style';
-import policer from '../../assets/images/policer.svg';
+import * as S from '../Chart.style';
+import policer from '../../../assets/images/policer.svg';
 
 const OverviewLargeCard = () => {
   return (
@@ -8,8 +8,8 @@ const OverviewLargeCard = () => {
       <S.InfoDiv>
         <S.TitleP>순찰 강화 필요 구역</S.TitleP>
         <S.StreetDiv>
-          <S.StreetP className='big'>정릉 3동</S.StreetP>
-          <S.StreetP className='small'>이번 달 기준</S.StreetP>
+          <S.StreetP className="big">정릉 3동</S.StreetP>
+          <S.StreetP className="small">이번 달 기준</S.StreetP>
         </S.StreetDiv>
       </S.InfoDiv>
     </S.OverviewLargeCardLayout>
